refactor(chat): 폴백 정책을 external-fallback 모듈로 분리 (파일 크기 가드)
external-provider.ts 가 폴백 래퍼 추가로 606줄이 되어 CI Gate 3(600줄)이
실패했다. 크램 대신 책임 경계로 자른다 — 스트리밍 루프 본체는 그대로 두고
재시도 정책(shouldFallbackToLocal + 진입 래퍼)만 새 모듈로 옮긴다.

- external-provider: runExternalStream 을 export (루프 본체), 530줄로 감소
- external-fallback: 폴백 정책 + dispatch 진입점. 소비처 편의를 위해
  ExternalProviderDeps/StreamFromExternalContext 타입 재노출
- ChatService 는 import 경로만 변경 (동작 무변화)

라이브 재검증: 분리 후 외부 모델 채팅 정상('파리'). 전체 2,158개 통과.
